rename drop_edge
