+++ tf_geometric/nn/__init__.py
@@ -11,6 +11,6 @@ from tf_geometric.nn.conv.sgc import *
 from tf_geometric.nn.conv.tagcn import *
 from tf_geometric.nn.conv.graph_sage import *
 from tf_geometric.nn.conv.appnp import *
-from tf_geometric.nn.sampling.dropedge import *
+from tf_geometric.nn.sampling.drop_edge import *
